160 & 163 - add support for multiple local lidar dataset (#164)

Mergining all changes as individual commit

+++ tests/test_many_stages_multiple_2/test_many_stages_multiple_2.py
@@ -43,9 +43,8 @@ class Test(unittest.TestCase):
         """Create a CatchmentGeometry object and then run the DemGenerator processing
         chain to download remote files and produce a DEM prior to testing."""
 
-        test_path = pathlib.Path().cwd() / pathlib.Path(
-            "tests/test_many_stages_multiple"
-        )
+        name = "test_many_stages_multiple_2"
+        test_path = pathlib.Path().cwd() / "tests" / name
 
         # Setup logging
         logging.basicConfig(
@@ -54,7 +53,7 @@ class Test(unittest.TestCase):
             level=logging.INFO,
             force=True,
         )
-        logging.info("In test_many_stages_multiple")
+        logging.info(f"In {name}")
 
         # load in the test instructions
         instruction_file_path = test_path / "instruction.json"
@@ -106,7 +105,8 @@ class Test(unittest.TestCase):
 
         # Cycle through all folders within the cache dir deleting their contents
         for path in cls.cache_dir.iterdir():
-            if path.is_dir():
+            # Ensure the local LiDAR dataset is not deleted.
+            if path.is_dir() and "Wellington_2013" not in str(path):
                 for file in path.glob("*"):  # only files
                     if file.is_file():
                         file.unlink()
@@ -143,6 +143,16 @@ class Test(unittest.TestCase):
             + "benchmark_dem",
         )
 
+        # compare the generated LiDAR data source maps
+        diff_array = test.lidar_source.data - benchmark.lidar_source.data
+        logging.info(f"DEM array diff is: {diff_array[diff_array != 0]}")
+        numpy.testing.assert_array_almost_equal(
+            test.lidar_source.data,
+            benchmark.lidar_source.data,
+            err_msg="The generated lidar_source test has different data from "
+            "the benchmark",
+        )
+
         # Compare the generated and benchmark roughness
         diff_array = (
             test.zo.data[~numpy.isnan(test.zo.data)]
@@ -152,8 +162,8 @@ class Test(unittest.TestCase):
         numpy.testing.assert_array_almost_equal(
             test.zo.data,
             benchmark.zo.data,
-            err_msg="The generated result_geofabric has different roughness data from "
-            "the benchmark_dem",
+            err_msg="The generated roughness test has different data from the "
+            "benchmark",
         )
 
         # explicitly free memory as xarray seems to be hanging onto memory
@@ -179,6 +189,7 @@ class Test(unittest.TestCase):
         )
         with rioxarray.rioxarray.open_rasterio(file_path, masked=True) as test:
             test.load()
+
         # compare the generated and benchmark DEMs
         diff_array = (
             test.z.data[~numpy.isnan(test.z.data)]
@@ -194,9 +205,20 @@ class Test(unittest.TestCase):
         )
         self.assertTrue(
             len(diff_array[numpy.abs(diff_array) > threshold]) < len(diff_array) / 250,
-            "More than 0.4% of DEM values differ by more than {threshold} on Linux test"
-            f" run: {diff_array[numpy.abs(diff_array) > threshold]} or "
-            f"{len(diff_array[numpy.abs(diff_array) > threshold]) / len(diff_array.flatten()) * 100}%",
+            f"More than 0.4% of DEM values differ by more than {threshold} on "
+            f"Linux test run: {diff_array[numpy.abs(diff_array) > threshold]} "
+            f"or {len(diff_array[numpy.abs(diff_array) > threshold]) / len(diff_array.flatten()) * 100}%",
+        )
+
+        # compare the generated and benchmark lidar source
+        diff_array = test.lidar_source.data - benchmark.lidar_source.data
+        numpy.testing.assert_array_almost_equal(
+            test.lidar_source.data,
+            benchmark.lidar_source.data,
+            decimal=3,
+            err_msg="The generated test has significantly different lidar "
+            "source values from the benchmark where there is LiDAR: "
+            f"{diff_array}",
         )
 
         # Compare the generated and benchmark roughnesses
